identifier location update
