+++ pyrit/identifiers/__init__.py
@@ -1,22 +1,23 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from pyrit.models.identifiers.class_name_utils import (
+from pyrit.identifiers.class_name_utils import (
     class_name_to_snake_case,
     snake_case_to_class_name,
 )
-from pyrit.models.identifiers.identifiers import (
-    Identifiable,
+from pyrit.identifiers.identifiers import (
     Identifier,
     IdentifierType,
 )
-from pyrit.models.identifiers.scorer_identifier import ScorerIdentifier
+from pyrit.identifiers.identifiable import Identifiable, LegacyIdentifiable
+from pyrit.identifiers.scorer_identifier import ScorerIdentifier
 
 __all__ = [
     "class_name_to_snake_case",
     "Identifiable",
     "Identifier",
     "IdentifierType",
+    "LegacyIdentifiable",
     "ScorerIdentifier",
     "snake_case_to_class_name",
 ]
