+++ pyrit/registry/class_registries/base_class_registry.py
@@ -20,7 +20,7 @@ from abc import ABC, abstractmethod
 from typing import Callable, Dict, Generic, Iterator, List, Optional, Type, TypeVar
 
 from pyrit.models import Identifier
-from pyrit.models.identifiers.class_name_utils import class_name_to_snake_case
+from pyrit.identifiers.class_name_utils import class_name_to_snake_case
 from pyrit.registry.base import RegistryProtocol
 
 # Type variable for the registered class type
